refactor: update package imports and improve command handling in Ralph Wiggum extension
- Changed import paths from "@mariozechner/pi-coding-agent" to "@earendil-works/pi-coding-agent" across multiple files.
- Consolidated command registration logic in the Ralph Wiggum extension for better maintainability.
- Enhanced task file handling to allow for existing content to be used in plan-level loops.
- Updated MCP configuration to include Costory command.
- Modified API key references in models.json to use environment variables.
- Updated default provider and model in settings.json for improved functionality.
