+++ home/.config/vscode/settings.json
@@ -102,16 +102,16 @@
 	"aws.telemetry": false,
 	"breadcrumbs.enabled": false,
 	"chat.instructionsFilesLocations": {
-		".claude/rules": true,
 		".github/instructions": true,
+		".claude/rules": true,
+		"~/.copilot/instructions": true,
+		"~/.claude/rules": true,
 		"/var/folders/jq/xkyvscnx7g5c9tn3dbmf09240000gq/T/postman-collections-post-response.instructions.md": true,
 		"/var/folders/jq/xkyvscnx7g5c9tn3dbmf09240000gq/T/postman-collections-pre-request.instructions.md": true,
 		"/var/folders/jq/xkyvscnx7g5c9tn3dbmf09240000gq/T/postman-folder-post-response.instructions.md": true,
 		"/var/folders/jq/xkyvscnx7g5c9tn3dbmf09240000gq/T/postman-folder-pre-request.instructions.md": true,
 		"/var/folders/jq/xkyvscnx7g5c9tn3dbmf09240000gq/T/postman-http-request-post-response.instructions.md": true,
-		"/var/folders/jq/xkyvscnx7g5c9tn3dbmf09240000gq/T/postman-http-request-pre-request.instructions.md": true,
-		"~/.claude/rules": true,
-		"~/.copilot/instructions": true
+		"/var/folders/jq/xkyvscnx7g5c9tn3dbmf09240000gq/T/postman-http-request-pre-request.instructions.md": true
 	},
 	"chat.mcp.gallery.enabled": true,
 	"chat.tools.terminal.autoApprove": {
@@ -326,7 +326,6 @@
 	"window.nativeTabs": true,
 	"window.restoreFullscreen": true,
 	"window.titleBarStyle": "native",
-	"workbench.colorTheme": "Dark Modern",
 	"workbench.editor.associations": {
 		"*.ipynb": "jupyter-notebook"
 	},
@@ -341,6 +340,8 @@
 	"workbench.layoutControl.enabled": false,
 	"workbench.panel.showLabels": false,
 	"workbench.secondarySideBar.defaultVisibility": "visible",
+	"workbench.startupEditor": "none",
 	"workbench.sideBar.location": "right",
-	"workbench.startupEditor": "none"
+	"chat.viewSessions.orientation": "stacked",
+	"workbench.colorTheme": "Dark Modern"
 }
