+++ home/.pi/agent/extensions/ralph-wiggum/loop-store.ts
@@ -5,7 +5,7 @@
 
 import * as fs from "node:fs";
 import * as path from "node:path";
-import type { ExtensionContext } from "@mariozechner/pi-coding-agent";
+import type { ExtensionContext } from "@earendil-works/pi-coding-agent";
 import { scratchDir, scratchDirFromFile, SCRATCH_DIR } from "./plan-paths";
 
 // ─── Types ───────────────────────────────────────────────────────────
