+++ home/.pi/agent/extensions/ralph-wiggum/plan-paths.ts
@@ -4,7 +4,7 @@
  */
 
 import * as path from "node:path";
-import type { ExtensionContext } from "@mariozechner/pi-coding-agent";
+import type { ExtensionContext } from "@earendil-works/pi-coding-agent";
 
 export const SCRATCH_DIR = ".scratch";
